etl/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.3
-r requirements_dev.txt